feat: paste in move mode
Share insert code with flyout insert.

Closes #478
